Add auto_annotate_models rake task
A new rake task was created to configure the defaults for the annotate gem. This task is set to run only in the development environment. It includes most of the options provided by the gem to customize its operation, ensuring that model annotations are done based on our application's needs.