Add HealthChecksController and route
A new controller, HealthChecksController, has been created to manage health checks. Additionally, a route named 'health_check' has been added that maps to the 'show' action of the new controller. This will respond with a simple 'ok' status, which can be used to verify the health and accessibility of the application.