Add 'show current user' functionality
This commit introduces a function to show the current user's account details. It includes the creation of two new classes: 'ShowCurrentUser' under both Operations and Serializers modules. In addition, an endpoint has been added to the UsersController. Minor adjustments have also been made to the UserAccount model and the routes configuration.